+++ src/components/Button/ButtonToggle/ButtonToggle.jsx
@@ -1,4 +1,4 @@
-import { BUTTON_TYPE } from '../../constants/button';
+import { BUTTON_TYPE } from '../../../constants/button';
 import styles from './ButtonToggle.module.scss';
 import { useState } from 'react';
