+++ src/components/Button/index.js
@@ -1,2 +1,2 @@
 export { default as Button } from './Button';
-export { default as ButtonToggle } from './ButtonToggle';
+export { default as ButtonToggle } from './ButtonToggle/ButtonToggle';
